requirements.txt
attrs==20.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
distro==1.5.0
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
Faker==8.1.0
idna==2.10
importlib-metadata==4.0.0
jsonschema==3.2.0
kafka-python==2.0.2
paramiko==2.7.2
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
text-unidecode==1.3
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.58.0
zipp==3.4.1
